[minimum-amount-aud] Merge containerisableComponents/helpers into helpers/contributions
